refactor(TechnicalMetricsDashboard): integrate useTechnicalMetrics hook and clean up code
